+++ docs/changelog.rst
@@ -12,6 +12,11 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.1.0/>`
 
 Each revision is versioned by the date of the revision.
 
+2026-04-17
+----------
+
+- Update documentation links to use intersphinx references.
+
 2026-04-16
 ----------